requirements.txt
streamlit
pandas
plotly.express
scikit-learn
train_test_split
StandardScaler
LabelEncoder
RandomForestClassifier
numpy